// src/app/technician/model/TechnicianApplicationStep.ts
export enum TechnicianApplicationStep {
  Guidance = 1,
  PersonalDetails = 2,
  Education = 3,
  EvidenceType = 4,
  WorkExperience = 5,
  PreviousApplicationsAndRegistrations = 6,
  SupportingDocuments = 7,
  Edi = 8,
  FtpDeclaration1 = 9,
  FtpDeclaration2 = 10,
  OverallDeclaration = 11,
  InitialReview = 12,
  Countersigning = 13,
  Review = 14,
  Payment = 15
}